Updated monetary fields to BigDecimal to save type conversion

+++ app/uk/gov/hmrc/lisaapi/models/ResorceModel.scala
@@ -96,11 +96,11 @@ case class LisaAccount(investorID:ID,
 
 case class TransferAccount(transferredFromAccountID:AccountId, transferredFromLMRN:ReferenceNumber, transferInDate:ISO8601Date)
 
-case class Bonuses(bonusDueForPeriod: Double, totalBonusDueYTD: Double, bonusPaidYTD: Option[Double], claimReason: String)
+case class Bonuses(bonusDueForPeriod: BigDecimal, totalBonusDueYTD: BigDecimal, bonusPaidYTD: Option[BigDecimal], claimReason: String)
 
-case class HelpToBuyTransfer(htbTransferInForPeriod: Double, htbTransferTotalYTD: Double)
+case class HelpToBuyTransfer(htbTransferInForPeriod: BigDecimal, htbTransferTotalYTD: BigDecimal)
 
-case class InboundPayments(newSubsForPeriod: Option[Double], newSubsYTD: Double, totalSubsForPeriod: Double, totalSubsYTD: Double)
+case class InboundPayments(newSubsForPeriod: Option[BigDecimal], newSubsYTD: BigDecimal, totalSubsForPeriod: BigDecimal, totalSubsYTD: BigDecimal)
 
 case class LifeEvent(accountID: AccountId,
                       lisaManagerReferenceNumber:ReferenceNumber,
